+++ src/muya/lib/parser/marked/lexer.js
@@ -43,6 +43,7 @@ Lexer.prototype.lex = function (src) {
 
 Lexer.prototype.token = function (src, top) {
   src = src.replace(/^ +$/gm, '')
+
   let loose
   let cap
   let bull
@@ -297,7 +298,6 @@ Lexer.prototype.token = function (src, top) {
         } else {
           prevItem = cap[i - 1]
         }
-
         // Determine whether item is loose or not. If previous item is loose
         // this item is also loose.
         loose = next = next || /^ *([*+-]|\d{1,9}(?:\.|\)))( +\S+\n\n(?!\s*$)|\n\n(?!\s*$))/.test(itemWithBullet)